changed default path, so it uses ORBIT_HOME rather than os.currentDir() as root directory
