+++ bintest/testFolder/templates/showverTemp
@@ -1,0 +1,7 @@
+Value Key (^\s+\S+)
+Value Value (.+$)
+
+Start
+  ^${Key} ${Value} -> Record
+
+EOF
